refactor(provider): move to signoz package
