+++ pkg/query-service/app/server.go
@@ -35,6 +35,8 @@ import (
 	"go.signoz.io/signoz/pkg/query-service/common"
 	"go.signoz.io/signoz/pkg/query-service/migrate"
 	v3 "go.signoz.io/signoz/pkg/query-service/model/v3"
+	"go.signoz.io/signoz/pkg/signoz"
+	"go.signoz.io/signoz/pkg/web"
 
 	"go.signoz.io/signoz/pkg/query-service/app/explorer"
 	"go.signoz.io/signoz/pkg/query-service/auth"
@@ -70,6 +72,7 @@ type ServerOptions struct {
 	Cluster           string
 	UseLogsNewSchema  bool
 	UseTraceNewSchema bool
+	SigNoz            *signoz.SigNoz
 }
 
 // Server runs HTTP, Mux and a grpc server
@@ -226,7 +229,7 @@ func NewServer(serverOptions *ServerOptions) (*Server, error) {
 		unavailableChannel: make(chan healthcheck.Status),
 	}
 
-	httpServer, err := s.createPublicServer(apiHandler)
+	httpServer, err := s.createPublicServer(apiHandler, serverOptions.SigNoz.Web)
 
 	if err != nil {
 		return nil, err
@@ -290,7 +293,7 @@ func (s *Server) createPrivateServer(api *APIHandler) (*http.Server, error) {
 	}, nil
 }
 
-func (s *Server) createPublicServer(api *APIHandler) (*http.Server, error) {
+func (s *Server) createPublicServer(api *APIHandler, web web.Web) (*http.Server, error) {
 
 	r := NewRouter()
 
@@ -334,6 +337,11 @@ func (s *Server) createPublicServer(api *APIHandler) (*http.Server, error) {
 	handler := c.Handler(r)
 
 	handler = handlers.CompressHandler(handler)
+
+	err := web.AddToRouter(r)
+	if err != nil {
+		return nil, err
+	}
 
 	return &http.Server{
 		Handler: handler,
