+++ pkg/query-service/main.go
@@ -9,11 +9,15 @@ import (
 	"time"
 
 	prommodel "github.com/prometheus/common/model"
+	"go.signoz.io/signoz/pkg/config"
+	"go.signoz.io/signoz/pkg/config/envprovider"
+	"go.signoz.io/signoz/pkg/config/fileprovider"
 	"go.signoz.io/signoz/pkg/query-service/app"
 	"go.signoz.io/signoz/pkg/query-service/auth"
 	"go.signoz.io/signoz/pkg/query-service/constants"
 	"go.signoz.io/signoz/pkg/query-service/migrate"
 	"go.signoz.io/signoz/pkg/query-service/version"
+	"go.signoz.io/signoz/pkg/signoz"
 
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
@@ -74,6 +78,22 @@ func main() {
 	logger := loggerMgr.Sugar()
 	version.PrintVersion()
 
+	config, err := signoz.NewConfig(context.Background(), config.ResolverConfig{
+		Uris: []string{"env:"},
+		ProviderFactories: []config.ProviderFactory{
+			envprovider.NewFactory(),
+			fileprovider.NewFactory(),
+		},
+	})
+	if err != nil {
+		zap.L().Fatal("Failed to create config", zap.Error(err))
+	}
+
+	signoz, err := signoz.New(context.Background(), config, signoz.NewProviderConfig())
+	if err != nil {
+		zap.L().Fatal("Failed to create signoz struct", zap.Error(err))
+	}
+
 	serverOptions := &app.ServerOptions{
 		HTTPHostPort:      constants.HTTPHostPort,
 		PromConfigPath:    promConfigPath,
@@ -90,6 +110,7 @@ func main() {
 		Cluster:           cluster,
 		UseLogsNewSchema:  useLogsNewSchema,
 		UseTraceNewSchema: useTraceNewSchema,
+		SigNoz:            signoz,
 	}
 
 	// Read the jwt secret key
